Add process_tiff script to run the whole pipeline

+++ deepflow_corrmatrix.py
@@ -11,7 +11,7 @@ from cvtools import readFlo
 
 import matplotlib.pyplot as plt
 import cv2 
-from flow import flow_err_deepflow
+from lib.flow import flow_err_deepflow
 
 def main():
 	usage = """deepflow_corrmatrix.py [register_dir]...
@@ -54,17 +54,16 @@ Ben Lansdell
 	for r1 in refframes:
 		for r2 in refframes:
 			if r1 != r2:
-
 				print("Comparing error in DeepFlow between frame %d and %d" %(r1, r2))
 				fn_in1 = args.dir_in + 'corrmatrix/%04d_%04d.flo'%(r1, r2)
 				fn_in2 = args.dir_in + 'corrmatrix/%04d_%04d.flo'%(r2, r1)
 
-				fwderr = flow_err_deepflow(fn_in1, fn_in2)
-				fwdtracked = fwderr < threshold
-
-				#Save error matrices
-				fn_out = args.dir_in + 'corrmatrix/%04d_%04d_deepflow_err.npz'%(r1, r2)
-				np.savez(fn_out, fwderr = fwderr, fwdtracked = fwdtracked)
+				if os.path.isfile(fn_in1) and os.path.isfile(fn_in2):
+					fwderr = flow_err_deepflow(fn_in1, fn_in2)
+					fwdtracked = fwderr < threshold
+					#Save error matrices
+					fn_out = args.dir_in + 'corrmatrix/%04d_%04d_deepflow_err.npz'%(r1, r2)
+					np.savez(fn_out, fwderr = fwderr, fwdtracked = fwdtracked)
 
 if __name__ == "__main__":
 	sys.exit(main())
